rename file movie_list_view to movie_list
